requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-generativeai==0.8.3
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.150.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
aiosqlite==0.20.0
apscheduler==3.10.4
httpx==0.27.2
pydantic-settings==2.6.1
langdetect==1.0.9
python-dotenv==1.0.1
twilio==9.3.7
alembic==1.14.0
